Fixes #17388. @petermattis.

I didn't trim down the data stored by LeaseHolderCache in this PR, I will try to fix this in another PR.

here is the benchmark test result for this PR:

```
before:
BenchmarkLeaseHolderCacheParallel-4   	  100000	     16587 ns/op
BenchmarkLeaseHolderCacheParallel-4   	  100000	     16618 ns/op
BenchmarkLeaseHolderCacheParallel-4   	  100000	     16684 ns/op
BenchmarkLeaseHolderCacheParallel-4   	  100000	     16589 ns/op
BenchmarkLeaseHolderCacheParallel-4   	  100000	     16416 ns/op

after:
BenchmarkLeaseHolderCacheParallel-4   	  200000	      7885 ns/op
BenchmarkLeaseHolderCacheParallel-4   	  200000	      9135 ns/op
BenchmarkLeaseHolderCacheParallel-4   	  100000	     10293 ns/op
BenchmarkLeaseHolderCacheParallel-4   	  200000	      9278 ns/op
BenchmarkLeaseHolderCacheParallel-4   	  200000	      7868 ns/op
```