@petermattis 2017-08-29
Approach, generally looks good, just a few small items below.

For comparing benchmarks, use [benchstat](https://godoc.org/golang.org/x/perf/cmd/benchstat). 

---

Review status: 0 of 2 files reviewed at latest revision, 5 unresolved discussions, all commit checks successful.

---

*[pkg/kv/leaseholder_cache.go, line 27 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17987#-KsiM7hVcJoLOJqGMx2m:-KsiM7hVcJoLOJqGMx2n:bsmej5i) ([raw file](https://github.com/cockroachdb/cockroach/blob/ff2426f7968aaabf72c163ba682536d9a00b743a/pkg/kv/leaseholder_cache.go#L27)):*
> ```Go
> 
> const (
> 	leaseHolderCacheShards = 16
> ```

I wonder if this should be a function of the number of cpus. 

---

*[pkg/kv/leaseholder_cache.go, line 53 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17987#-KsiLseEDZcwJsxgCkVN:-KsiLseEDZcwJsxgCkVO:betlu6t) ([raw file](https://github.com/cockroachdb/cockroach/blob/ff2426f7968aaabf72c163ba682536d9a00b743a/pkg/kv/leaseholder_cache.go#L53)):*
> ```Go
> 			Policy: cache.CacheLRU,
> 			ShouldEvict: func(s int, key, value interface{}) bool {
> 				return int64(s) > size()
> ```

I think this should be `int64(s) > size()/int64(len(leaseholderCache.shards)))`. As it is currently you've made the total leaseholder cache size 16x larger. 

---

*[pkg/kv/leaseholder_cache.go, line 90 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17987#-KsiMEYoikIbkGVnkLXb:-KsiMEYoikIbkGVnkLXc:bm06u37) ([raw file](https://github.com/cockroachdb/cockroach/blob/ff2426f7968aaabf72c163ba682536d9a00b743a/pkg/kv/leaseholder_cache.go#L90)):*
> ```Go
> 	if (repDesc == roachpb.ReplicaDescriptor{}) {
> 		if log.V(2) {
> 			log.Infof(ctx, "evicting lease holder for r%d", rangeID)
> ```

Is there a reason for this change? We generally prefix the range ID on log messages and this is now inconsistent with other log messages in this file. Also, `leaseholder` is one word. 

---

*[pkg/kv/leaseholder_cache_test.go, line 81 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17987#-KsiMJ6a4EKqjAWE94cV:-KsiMJ6a4EKqjAWE94cW:b-cbeajs) ([raw file](https://github.com/cockroachdb/cockroach/blob/ff2426f7968aaabf72c163ba682536d9a00b743a/pkg/kv/leaseholder_cache_test.go#L81)):*
> ```Go
> 	b.RunParallel(func(pb *testing.PB) {
> 		for pb.Next() {
> 			for i := 1; i < 1<<6; i++ {
> ```

Idiomatic benchmarks execute a single iteration in the loop. I suggest doing something like:

```
  const numRanges = 1<<6
  ...
  var n int64
  for pb.Next() {
    rangeID := roachpb.RangeID(n % numRanges)
    n++
    if _, ok := lc.Lookup(ctx, rangeID); !ok {
      ...
    }
  }
```

---

*[pkg/kv/leaseholder_cache_test.go, line 83 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17987#-KsiMfT8WfQKz3mih1nW:-KsiMfT8WfQKz3mih1nX:b-mylyqk) ([raw file](https://github.com/cockroachdb/cockroach/blob/ff2426f7968aaabf72c163ba682536d9a00b743a/pkg/kv/leaseholder_cache_test.go#L83)):*
> ```Go
> 			for i := 1; i < 1<<6; i++ {
> 				rangeID := roachpb.RangeID(i)
> 				replica := roachpb.ReplicaDescriptor{StoreID: roachpb.StoreID(i)}
> ```

This is only being used for logging. I'd get rid of it.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/17987#-:-KsiLhvK39BUR9OciSH9:b-p2xy4)*
<!-- Sent from Reviewable.io -->