@bdarnell 2017-08-29
Reviewed 2 of 2 files at r1.
Review status: all files reviewed at latest revision, 5 unresolved discussions, all commit checks successful.

---

*[pkg/kv/leaseholder_cache.go, line 90 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17987#-KsiMEYoikIbkGVnkLXb:-KsiQTaJpxHuCvUFO3OE:b-4e20yb) ([raw file](https://github.com/cockroachdb/cockroach/blob/ff2426f7968aaabf72c163ba682536d9a00b743a/pkg/kv/leaseholder_cache.go#L90)):*
> Also, leaseholder is one word.

But we capitalize the H in the type name `LeaseHolderCache` as if it were two words. 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/17987)*
<!-- Sent from Reviewable.io -->