@a6802739 2017-08-31
here is the result :
```
before:
BenchmarkLeaseHolderCacheParallel-4      5000000               279 ns/op
BenchmarkLeaseHolderCacheParallel-4      5000000               297 ns/op
BenchmarkLeaseHolderCacheParallel-4      5000000               287 ns/op
BenchmarkLeaseHolderCacheParallel-4      5000000               284 ns/op
BenchmarkLeaseHolderCacheParallel-4      5000000               292 ns/op

after:
BenchmarkLeaseHolderCacheParallel-4     10000000               175 ns/op
BenchmarkLeaseHolderCacheParallel-4     10000000               169 ns/op
BenchmarkLeaseHolderCacheParallel-4     10000000               172 ns/op
BenchmarkLeaseHolderCacheParallel-4     10000000               190 ns/op
BenchmarkLeaseHolderCacheParallel-4     10000000               176 ns/op
```


here is the `benchstat` result:
```
name                        old time/op  new time/op  delta
LeaseHolderCacheParallel-4   288ns ± 3%   176ns ± 8%  -38.71%  (p=0.008 n=5+5)
```

---

Review status: 0 of 2 files reviewed at latest revision, 5 unresolved discussions, all commit checks successful.

---

*[pkg/kv/leaseholder_cache.go, line 27 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17987#-KsiM7hVcJoLOJqGMx2m:-Ksr7bWlDSYEB-do6Hs6:b30r698) ([raw file](https://github.com/cockroachdb/cockroach/blob/ff2426f7968aaabf72c163ba682536d9a00b743a/pkg/kv/leaseholder_cache.go#L27)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary><blockquote>

I wonder if this should be a function of the number of cpus. 
</blockquote></details>

I have done this now. But I'm not sure it is appropriate.

---

*[pkg/kv/leaseholder_cache.go, line 53 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17987#-KsiLseEDZcwJsxgCkVN:-KsqyvIFUPAf9fZkGDoI:b7j0jo8) ([raw file](https://github.com/cockroachdb/cockroach/blob/ff2426f7968aaabf72c163ba682536d9a00b743a/pkg/kv/leaseholder_cache.go#L53)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary><blockquote>

I think this should be `int64(s) > size()/int64(len(leaseholderCache.shards)))`. As it is currently you've made the total leaseholder cache size 16x larger. 
</blockquote></details>

Yeah, My mistake.

---

*[pkg/kv/leaseholder_cache.go, line 90 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17987#-KsiMEYoikIbkGVnkLXb:-Ksr8E2sO--fhhfs41h_:bsdcgy6) ([raw file](https://github.com/cockroachdb/cockroach/blob/ff2426f7968aaabf72c163ba682536d9a00b743a/pkg/kv/leaseholder_cache.go#L90)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary><blockquote>

Yeah, we should fix that. I'm pretty sure we've standardized on `leaseholder` and `Leaseholder` elsewhere.
</blockquote></details>

Because my branch is far behind the master branch. Fixed it now.

---

*[pkg/kv/leaseholder_cache_test.go, line 81 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17987#-KsiMJ6a4EKqjAWE94cV:-Ksr8UxmIpUMj4m4F9_1:boyexbg) ([raw file](https://github.com/cockroachdb/cockroach/blob/ff2426f7968aaabf72c163ba682536d9a00b743a/pkg/kv/leaseholder_cache_test.go#L81)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary><blockquote>

Idiomatic benchmarks execute a single iteration in the loop. I suggest doing something like:

```
  const numRanges = 1<<6
  ...
  var n int64
  for pb.Next() {
    rangeID := roachpb.RangeID(n % numRanges)
    n++
    if _, ok := lc.Lookup(ctx, rangeID); !ok {
      ...
    }
  }
```
</blockquote></details>

Done. Thanks.

---

*[pkg/kv/leaseholder_cache_test.go, line 83 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17987#-KsiMfT8WfQKz3mih1nW:-Ksr8c-TTPxZSpmn6pob:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/ff2426f7968aaabf72c163ba682536d9a00b743a/pkg/kv/leaseholder_cache_test.go#L83)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary><blockquote>

This is only being used for logging. I'd get rid of it.
</blockquote></details>

Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/17987#-:-Ksr8d8V5TmiE39r7s0Q:brvuor)*
<!-- Sent from Reviewable.io -->