@a6802739 2017-09-04
@petermattis 

here is the newest result:

```
before:
before:
BenchmarkLeaseHolderCacheParallel-4      5000000               277 ns/op
BenchmarkLeaseHolderCacheParallel-4      5000000               275 ns/op
BenchmarkLeaseHolderCacheParallel-4      5000000               273 ns/op
BenchmarkLeaseHolderCacheParallel-4      5000000               276 ns/op
BenchmarkLeaseHolderCacheParallel-4      5000000               281 ns/op

after:
after:
BenchmarkLeaseHolderCacheParallel-4     10000000               121 ns/op
BenchmarkLeaseHolderCacheParallel-4     10000000               142 ns/op
BenchmarkLeaseHolderCacheParallel-4     10000000               158 ns/op
BenchmarkLeaseHolderCacheParallel-4     10000000               152 ns/op
BenchmarkLeaseHolderCacheParallel-4     10000000               129 ns/op

name                        old time/op  new time/op  delta
LeaseHolderCacheParallel-4   276ns ± 2%   140ns ±14%  -49.20%  (p=0.008 n=5+5)
```

I don't quite understand    `Rather that accessing staticShards here, I'd move this down a line and change it to: numRanges := 2 * len(lc.shards). Either way you're making some assumption about the implementation of leaseholder cache, and the latter way seems more explicit.`

Thanks for your kind review.